###fix Modified delete method (Lessor and Property )

+++ Habitarium-Gitflow-Maven/src/main/java/dao/LessorDAO.java
@@ -56,9 +56,14 @@ public class LessorDAO implements DAO<Lessor> {
         Lessor lessor = null;
         try {
             lessor = entityManager.find(Lessor.class, id);
-            this.entityManager.getTransaction().begin();
-            this.entityManager.remove(lessor);
-            this.entityManager.getTransaction().commit();
+            if(lessor.getRent() == null){
+                this.entityManager.getTransaction().begin();
+                this.entityManager.remove(lessor);
+                this.entityManager.getTransaction().commit();
+            } else {
+                System.out.println("Nao eh possivel apagar um locatario vinculado a um aliguel");
+            }
+
         } catch (Exception exception) {
             this.entityManager.getTransaction().rollback();
         } finally {
