+++ Habitarium-Gitflow-Maven/src/main/java/dao/PropertyDAO.java
@@ -55,9 +55,14 @@ public class PropertyDAO implements DAO<Property> {
         Property property = null;
         try {
             property = entityManager.find(Property.class, id);
-            this.entityManager.getTransaction().begin();
-            this.entityManager.remove(property);
-            this.entityManager.getTransaction().commit();
+            if(property.getRent() == null){
+                this.entityManager.getTransaction().begin();
+                this.entityManager.remove(property);
+                this.entityManager.getTransaction().commit();
+            } else {
+                System.out.println("Nao eh possivel apagar uma propriedade vinculada a um aluguel");
+            }
+
         } catch (Exception exception) {
             this.entityManager.getTransaction().rollback();
         } finally {
